issue9 코드 리팩토링  - User api 단건/다건 api 통합  - mapper sql 코드 가독성 포맷 수정  - 회원 가입시 로그인 aop 적용부분 삭제  - 불필요한 주석삭제  - entity와 dto>request class 분리  - xml 코드 규칙에 통일 시켜 정리
